+++ turtlebot3_dqn/turtlebot3_dqn/dqn_gazebo.py
@@ -104,7 +104,7 @@ class GazeboInterface(Node):
                 self.get_logger().warn('service for spawn_entity is not available, waiting ...')
             future = self.spawn_entity_client.call_async(spawn_req)
             rclpy.spin_until_future_complete(self, future)
-            print(f'[✓] Spawn Goal at ({self.entity_pose_x}, {self.entity_pose_y}, {0.0})')
+            print(f'Spawn Goal at ({self.entity_pose_x}, {self.entity_pose_y}, {0.0})')
         else:
             service_name = '/world/dqn/create'
             package_share = get_package_share_directory('turtlebot3_gazebo')
@@ -129,7 +129,7 @@ class GazeboInterface(Node):
             ]
             try:
                 subprocess.run(cmd, check=True, stdout=subprocess.DEVNULL)
-                print(f'[✓] Spawn Goal at ({self.entity_pose_x}, {self.entity_pose_y}, {0.0})')
+                print(f'Spawn Goal at ({self.entity_pose_x}, {self.entity_pose_y}, {0.0})')
             except subprocess.CalledProcessError:
                 pass
 
@@ -142,7 +142,7 @@ class GazeboInterface(Node):
                 self.get_logger().warn('service for delete_entity is not available, waiting ...')
             future = self.delete_entity_client.call_async(delete_req)
             rclpy.spin_until_future_complete(self, future)
-            print('[✓] Delete Goal')
+            print('Delete Goal')
         else:
             service_name = '/world/dqn/remove'
             req = f'name: "{self.entity_name}", type: 2'
@@ -156,7 +156,7 @@ class GazeboInterface(Node):
             ]
             try:
                 subprocess.run(cmd, check=True, stdout=subprocess.DEVNULL)
-                print('[✓] Delete Goal')
+                print('Delete Goal')
             except subprocess.CalledProcessError:
                 pass
 
@@ -181,7 +181,7 @@ class GazeboInterface(Node):
         ]
         try:
             subprocess.run(cmd_delete, check=True, stdout=subprocess.DEVNULL)
-            print('[✓] Delete Burger')
+            print('Delete Burger')
         except subprocess.CalledProcessError:
             pass
         time.sleep(0.2)
@@ -203,7 +203,7 @@ class GazeboInterface(Node):
         ]
         try:
             subprocess.run(cmd_spawn, check=True, stdout=subprocess.DEVNULL)
-            print('[✓] Spawn Burger')
+            print('Spawn Burger')
         except subprocess.CalledProcessError:
             pass