+++ turtlebot3_machine_learning/CHANGELOG.rst
@@ -5,11 +5,11 @@ Changelog for package turtlebot3_machine_learning
 ------------------
 * Support for ROS 2 Jazzy version
 * Using Gazebo sim for package
-Contributors: ChanHyeong Lee
+* Contributors: ChanHyeong Lee
 
 1.0.0 (2025-04-17)
 ------------------
 * Support for ROS 2 Humble version
 * Renewal of package structure
 * Improved behavioral rewards for agents
-Contributors: ChanHyeong Lee
+* Contributors: ChanHyeong Lee