+++ turtlebot3_machine_learning/CHANGELOG.rst
@@ -1,10 +1,11 @@
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 Changelog for package turtlebot3_machine_learning
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
 1.0.1 (2025-05-02)
 ------------------
 * Support for ROS 2 Jazzy version
-* Using Gazebo sim for package
+* Gazebo simulation support for the package
 * Contributors: ChanHyeong Lee
 
 1.0.0 (2025-04-17)