+++ turtlebot3_machine_learning/CHANGELOG.rst
@@ -2,9 +2,15 @@
 Changelog for package turtlebot3_machine_learning
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+1.0.1 (2025-05-02)
+------------------
+* Support for ROS 2 Jazzy version
+* Gazebo simulation support for the package
+* Contributors: ChanHyeong Lee
+
 1.0.0 (2025-04-17)
 ------------------
 * Support for ROS 2 Humble version
 * Renewal of package structure
 * Improved behavioral rewards for agents
-Contributors: ChanHyeong Lee
+* Contributors: ChanHyeong Lee